requirements.txt
thordata-sdk>=1.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
colorama>=0.4.6